--- 15.10.21 ttc/15.10.21 ttc/Debug/15.10.21 ttc.tlog/15.10.21 ttc.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.29.30133:TargetPlatformVersion=10.0.19041.0:
-Debug|Win32|D:\c++\Informatics-and-Programming\15.10.21 ttc\|
