--- 15.10.21 ttc/15.10.21 ttc/Debug/15.10.21 ttc.tlog/CL.command.1.tlog
@@ -1,1 +1,0 @@
-ÿþ
