+++ 15.10.21ttc/15.10.21 ttc.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "15.10.21 ttc", "15.10.21 ttc\15.10.21 ttc.vcxproj", "{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Debug|x64.ActiveCfg = Debug|x64
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Debug|x64.Build.0 = Debug|x64
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Debug|x86.ActiveCfg = Debug|Win32
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Debug|x86.Build.0 = Debug|Win32
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Release|x64.ActiveCfg = Release|x64
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Release|x64.Build.0 = Release|x64
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Release|x86.ActiveCfg = Release|Win32
+		{E54A2550-AA0C-4832-9AC5-4257F0FBE00E}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2C10DDD7-D130-4517-82ED-FF6DC09A15C2}
+	EndGlobalSection
+EndGlobal
