+++ 15.10.21ttc/15.10.21 ttc/Debug/15.10.21 ttc.log
@@ -1,0 +1,3 @@
+﻿  Source.cpp
+D:\c++\Informatics-and-Programming\15.10.21 ttc\15.10.21 ttc\Source.cpp(18,8): warning C4101: max: неиспользованная локальная переменная
+  15.10.21 ttc.vcxproj -> D:\c++\Informatics-and-Programming\15.10.21 ttc\Debug\15.10.21 ttc.exe
